feat: rely on the Component API
