+++ bundlesize.config.json
@@ -26,7 +26,7 @@
     },
     {
       "path": "packages/autocomplete-layout-classic/dist/umd/index.production.js",
-      "maxSize": "4 kB"
+      "maxSize": "4.25 kB"
     },
     {
       "path": "packages/autocomplete-theme-classic/dist/theme.css",
