--- examples/js/app.tsx
@@ -1,167 +1,0 @@
-/** @jsx h */
-import {
-  autocomplete,
-  getAlgoliaHits,
-  snippetHit,
-} from '@algolia/autocomplete-js';
-import {
-  AutocompleteInsightsApi,
-  createAlgoliaInsightsPlugin,
-} from '@algolia/autocomplete-plugin-algolia-insights';
-import { createQuerySuggestionsPlugin } from '@algolia/autocomplete-plugin-query-suggestions';
-import { createLocalStorageRecentSearchesPlugin } from '@algolia/autocomplete-plugin-recent-searches';
-import algoliasearch from 'algoliasearch';
-import { h, Fragment } from 'preact';
-import insightsClient from 'search-insights';
-
-import '@algolia/autocomplete-theme-classic';
-
-import { createCategoriesPlugin } from './categoriesPlugin';
-import { shortcutsPlugin } from './shortcutsPlugin';
-import { ProductHit, ProductRecord } from './types';
-
-const appId = 'latency';
-const apiKey = '6be0576ff61c053d5f9a3225e2a90f76';
-const searchClient = algoliasearch(appId, apiKey);
-insightsClient('init', { appId, apiKey });
-
-const algoliaInsightsPlugin = createAlgoliaInsightsPlugin({ insightsClient });
-const recentSearchesPlugin = createLocalStorageRecentSearchesPlugin({
-  key: 'search',
-  limit: 3,
-});
-const querySuggestionsPlugin = createQuerySuggestionsPlugin({
-  searchClient,
-  indexName: 'instant_search_demo_query_suggestions',
-  getSearchParams({ state }) {
-    return recentSearchesPlugin.data.getAlgoliaSearchParams({
-      clickAnalytics: true,
-      hitsPerPage: state.query ? 5 : 10,
-    });
-  },
-  categoryAttribute: [
-    'instant_search',
-    'facets',
-    'exact_matches',
-    'categories',
-  ],
-});
-const categoriesPlugin = createCategoriesPlugin({ searchClient });
-
-autocomplete({
-  container: '#autocomplete',
-  placeholder: 'Search',
-  debug: true,
-  openOnFocus: true,
-  plugins: [
-    shortcutsPlugin,
-    algoliaInsightsPlugin,
-    recentSearchesPlugin,
-    querySuggestionsPlugin,
-    categoriesPlugin,
-  ],
-  getSources({ query, state }) {
-    if (!query) {
-      return [];
-    }
-
-    return [
-      {
-        sourceId: 'products',
-        getItems() {
-          return getAlgoliaHits<ProductRecord>({
-            searchClient,
-            queries: [
-              {
-                indexName: 'instant_search',
-                query,
-                params: {
-                  clickAnalytics: true,
-                  attributesToSnippet: ['name:10', 'description:35'],
-                  snippetEllipsisText: '…',
-                },
-              },
-            ],
-          });
-        },
-        templates: {
-          header() {
-            return (
-              <Fragment>
-                <span className="aa-SourceHeaderTitle">Products</span>
-                <div className="aa-SourceHeaderLine" />
-              </Fragment>
-            );
-          },
-          item({ item }) {
-            return (
-              <ProductItem
-                hit={item}
-                insights={state.context.algoliaInsightsPlugin.insights}
-              />
-            );
-          },
-          noResults() {
-            return (
-              <div className="aa-ItemContent">No products for this query.</div>
-            );
-          },
-        },
-      },
-    ];
-  },
-});
-
-type ProductItemProps = {
-  hit: ProductHit;
-  insights: AutocompleteInsightsApi;
-};
-
-function ProductItem({ hit, insights }: ProductItemProps) {
-  return (
-    <Fragment>
-      <div className="aa-ItemIcon aa-ItemIcon--alignTop">
-        <img src={hit.image} alt={hit.name} width="40" height="40" />
-      </div>
-      <div className="aa-ItemContent">
-        <div className="aa-ItemContentTitle">
-          {snippetHit<ProductHit>({ hit, attribute: 'name' })}
-        </div>
-        <div className="aa-ItemContentDescription">
-          {snippetHit<ProductHit>({ hit, attribute: 'description' })}
-        </div>
-      </div>
-      <div className="aa-ItemActions">
-        <button
-          className="aa-ItemActionButton aa-TouchOnly aa-ActiveOnly"
-          type="button"
-          title="Select"
-        >
-          <svg viewBox="0 0 24 24" width="20" height="20" fill="currentColor">
-            <path d="M18.984 6.984h2.016v6h-15.188l3.609 3.609-1.406 1.406-6-6 6-6 1.406 1.406-3.609 3.609h13.172v-4.031z" />
-          </svg>
-        </button>
-        <button
-          className="aa-ItemActionButton"
-          type="button"
-          title="Add to cart"
-          onClick={(event) => {
-            event.preventDefault();
-            event.stopPropagation();
-
-            insights.convertedObjectIDsAfterSearch({
-              eventName: 'Added to cart',
-              index: hit.__autocomplete_indexName,
-              objectIDs: [hit.objectID],
-              queryID: hit.__autocomplete_queryID,
-            });
-          }}
-        >
-          <svg viewBox="0 0 24 24" width="18" height="18" fill="currentColor">
-            <path d="M19 5h-14l1.5-2h11zM21.794 5.392l-2.994-3.992c-0.196-0.261-0.494-0.399-0.8-0.4h-12c-0.326 0-0.616 0.156-0.8 0.4l-2.994 3.992c-0.043 0.056-0.081 0.117-0.111 0.182-0.065 0.137-0.096 0.283-0.095 0.426v14c0 0.828 0.337 1.58 0.879 2.121s1.293 0.879 2.121 0.879h14c0.828 0 1.58-0.337 2.121-0.879s0.879-1.293 0.879-2.121v-14c0-0.219-0.071-0.422-0.189-0.585-0.004-0.005-0.007-0.010-0.011-0.015zM4 7h16v13c0 0.276-0.111 0.525-0.293 0.707s-0.431 0.293-0.707 0.293h-14c-0.276 0-0.525-0.111-0.707-0.293s-0.293-0.431-0.293-0.707zM15 10c0 0.829-0.335 1.577-0.879 2.121s-1.292 0.879-2.121 0.879-1.577-0.335-2.121-0.879-0.879-1.292-0.879-2.121c0-0.552-0.448-1-1-1s-1 0.448-1 1c0 1.38 0.561 2.632 1.464 3.536s2.156 1.464 3.536 1.464 2.632-0.561 3.536-1.464 1.464-2.156 1.464-3.536c0-0.552-0.448-1-1-1s-1 0.448-1 1z" />
-          </svg>
-        </button>
-      </div>
-    </Fragment>
-  );
-}
