+++ packages/autocomplete-layout-classic/package.json
@@ -30,6 +30,9 @@
     "prepare": "yarn build:esm && yarn build:types",
     "watch": "watch \"yarn on:change\" --ignoreDirectoryPattern \"/dist/\""
   },
+  "dependencies": {
+    "preact": "^10.0.0"
+  },
   "devDependencies": {
     "@algolia/autocomplete-js": "1.0.0-alpha.44"
   }
