+++ packages/autocomplete-layout-classic/src/NavigationCommands.d.ts
@@ -1,0 +1,7 @@
+import { NavigationCommandsProps } from './NavigationCommands';
+
+declare module '@algolia/autocomplete-js' {
+  export interface AutocompleteComponents {
+    NavigationCommands: (props: NavigationCommandsProps) => JSX.Element;
+  }
+}
