+++ packages/autocomplete-layout-classic/src/SearchByAlgolia.d.ts
@@ -1,0 +1,7 @@
+import { SearchByAlgoliaProps } from './SearchByAlgolia';
+
+declare module '@algolia/autocomplete-js' {
+  export interface AutocompleteComponents {
+    SearchByAlgolia: (props: SearchByAlgoliaProps) => JSX.Element;
+  }
+}
